+++ SlackFilter/ServiceClients/Model/BuildDefinition.cs
@@ -1,9 +1,10 @@
-﻿namespace SlackFilter.ServiceClients
+﻿namespace SlackFilter.ServiceClients.Model
 {
-    public class ReleaseDefinition
+    public class BuildDefinition
     {
         public int Id { get; set; }
         public string Name { get; set; }
         public string Path { get; set; }
+        public string QueueStatus { get; set; }
     }
 }
