+++ SlackFilter/ServiceClients/Model/GetBuildDefinitionResult.cs
@@ -1,4 +1,4 @@
-﻿namespace SlackFilter.ServiceClients
+﻿namespace SlackFilter.ServiceClients.Model
 {
     public class GetBuildDefinitionResult
     {
