+++ SlackFilter/ServiceClients/Model/ReleaseDefinition.cs
@@ -1,10 +1,9 @@
-﻿namespace SlackFilter.ServiceClients
+﻿namespace SlackFilter.ServiceClients.Model
 {
-    public class BuildDefinition
+    public class ReleaseDefinition
     {
         public int Id { get; set; }
         public string Name { get; set; }
         public string Path { get; set; }
-        public string QueueStatus { get; set; }
     }
 }
